% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uaiscore.R
\name{uaiscore}
\alias{uaiscore}
\title{The function uaiscore is designed for predictive modeling of gene expression data.}
\usage{
uaiscore(
  eset,
  pdata = NULL,
  id_pdata = "ID",
  scale = FALSE,
  check_eset = FALSE
)
}
\arguments{
\item{eset}{Required matrix or data frame containing the gene expression data to be used for prediction. If not a matrix, it will be coerced into one.}

\item{pdata}{An optional data frame containing sample data to be used during gene expression prediction. Defaults to NULL.}

\item{id_pdata}{An optional character vector indicating the name of the sample ID column to be used in pdata data frame. Defaults to "ID".}

\item{scale}{An optional logical value indicating whether to scale the gene expression data. Default value is FALSE.}

\item{check_eset}{An optional logical value indicating whether to remove outlier genes from the gene expression data. Default value is FALSE.}
}
\description{
The function uaiscore is designed for predictive modeling of gene expression data.
}
\author{
Dongqiang Zeng
}
